requirements.txt
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
future==1.0.0
h5py==3.14.0
hf-xet==1.1.8
huggingface-hub==0.34.4
hyperopt==0.2.7
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lightning-utilities==0.15.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
nibabel==5.3.2
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
py4j==0.10.9.9
pyaml==25.7.0
pyarrow==21.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1731006091482/work
pyzstd==0.17.0
ray==2.48.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
safetensors==0.6.2
scikit-learn==1.7.1
scikit-optimize==0.10.2
scipy==1.16.1
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.19
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.8.0
torchsummary==1.5.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
